Title: 가능한 수열 중 최솟값 구하기, Time: 125ms, Memory: 10MB, Status: Runtime Error - Codetree

+++ 250821/가능한 수열 중 최솟값 구하기/find-min-of-possible-series.java
@@ -2,7 +2,7 @@ import java.util.*;
 
 public class Main {
 
-    static int min = Long.MAX_VALUE;
+    static long min = Long.MAX_VALUE;
     static int n;
     static List<Integer> list = new ArrayList<>();
     public static void main(String[] args) {
